Alternate mounting bracket
Designed to bend around the X-motor pulley case and hold the daughterboard at 90 degrees to the carriage plate.  Intended to be used to relocate the OEM daughterboard and its plastic case/LED holder. Some cables may be too short to use this mount if you discard the OEM case, and the LED will not illuminate the nozzle any more.  Unless you have a BTT board, which signals "firmware flashing in process" with the LED, you may just want to remove the LED.